+++ pcs/cli/common/errors.py
@@ -1,30 +1,16 @@
-from typing import (
-    List,
-    Optional,
-)
+from typing import Optional
 
 from pcs.common.str_tools import (
     format_list_base,
     quote_items,
 )
+from pcs.common.types import StringSequence
 
 ERR_NODE_LIST_AND_ALL_MUTUALLY_EXCLUSIVE = (
     "Cannot specify both --all and a list of nodes."
 )
 
 SEE_MAN_CHANGES = "See 'man pcs' -> Changes in pcs-{}."
-
-
-def _msg_command_replaced(new_commands: List[str], pcs_version: str) -> str:
-    commands = format_list_base(quote_items(new_commands))
-    changes = SEE_MAN_CHANGES.format(pcs_version)
-    return f"This command has been replaced with {commands}. {changes}"
-
-
-def raise_command_replaced(new_commands: List[str], pcs_version: str) -> None:
-    raise CmdLineInputError(
-        message=_msg_command_replaced(new_commands, pcs_version=pcs_version)
-    )
 
 
 class CmdLineInputError(Exception):
@@ -53,3 +39,25 @@ class CmdLineInputError(Exception):
         self.message = message
         self.hint = hint
         self.show_both_usage_and_message = show_both_usage_and_message
+
+
+def _msg_command_replaced(
+    new_commands: StringSequence, pcs_version: str
+) -> str:
+    commands = format_list_base(quote_items(new_commands))
+    changes = SEE_MAN_CHANGES.format(pcs_version)
+    return f"This command has been replaced with {commands}. {changes}"
+
+
+def command_replaced(
+    new_commands: StringSequence, pcs_version: str
+) -> CmdLineInputError:
+    return CmdLineInputError(
+        message=_msg_command_replaced(new_commands, pcs_version=pcs_version)
+    )
+
+
+def raise_command_replaced(
+    new_commands: StringSequence, pcs_version: str
+) -> None:
+    raise command_replaced(new_commands, pcs_version)