+++ pcs/cli/routing/resource_stonith_common.py
@@ -4,11 +4,42 @@ from typing import (
 )
 
 from pcs import resource
-from pcs.cli.common.parse_args import InputModifiers
+from pcs.cli.common.errors import (
+    command_replaced,
+    raise_command_replaced,
+)
+from pcs.cli.common.parse_args import (
+    Argv,
+    InputModifiers,
+)
 from pcs.cli.common.routing import (
     CliCmdInterface,
     create_router,
 )
+
+
+def resource_show(
+    lib: Any, argv: Argv, modifiers: InputModifiers, stonith: bool = False
+) -> None:
+    """
+    Options:
+      * -f - CIB file
+      * --full - print all configured options
+      * --groups - print resource groups
+      * --hide-inactive - print only active resources
+    """
+    del lib
+    modifiers.ensure_only_supported(
+        "-f", "--full", "--groups", "--hide-inactive"
+    )
+    if modifiers.get("--groups"):
+        raise_command_replaced(["pcs resource group list"], pcs_version="0.11")
+
+    keyword = "stonith" if stonith else "resource"
+    if modifiers.get("--full") or argv:
+        raise_command_replaced([f"pcs {keyword} config"], pcs_version="0.11")
+
+    raise_command_replaced([f"pcs {keyword} status"], pcs_version="0.11")
 
 
 def resource_defaults_cmd(parent_cmd: List[str]) -> CliCmdInterface:
@@ -21,9 +52,8 @@ def resource_defaults_cmd(parent_cmd: List[str]) -> CliCmdInterface:
           * --force - allow unknown options
         """
         if argv and "=" in argv[0]:
-            # DEPRECATED legacy command
-            return resource.resource_defaults_legacy_cmd(
-                lib, argv, modifiers, deprecated_syntax_used=True
+            raise command_replaced(
+                ["pcs resource defaults update"], pcs_version="0.12"
             )
 
         router = create_router(
@@ -38,7 +68,7 @@ def resource_defaults_cmd(parent_cmd: List[str]) -> CliCmdInterface:
                     },
                     parent_cmd + ["set"],
                 ),
-                "update": resource.resource_defaults_legacy_cmd,
+                "update": resource.resource_defaults_update_cmd,
             },
             parent_cmd,
             default_cmd="config",
@@ -58,9 +88,8 @@ def resource_op_defaults_cmd(parent_cmd: List[str]) -> CliCmdInterface:
           * --force - allow unknown options
         """
         if argv and "=" in argv[0]:
-            # DEPRECATED legacy command
-            return resource.resource_op_defaults_legacy_cmd(
-                lib, argv, modifiers, deprecated_syntax_used=True
+            raise command_replaced(
+                ["pcs resource op defaults update"], pcs_version="0.12"
             )
 
         router = create_router(
@@ -75,7 +104,7 @@ def resource_op_defaults_cmd(parent_cmd: List[str]) -> CliCmdInterface:
                     },
                     parent_cmd + ["set"],
                 ),
-                "update": resource.resource_op_defaults_legacy_cmd,
+                "update": resource.resource_op_defaults_update_cmd,
             },
             parent_cmd,
             default_cmd="config",